Initial Graphics
Added the initial graphics, which is not much, and it's not pretty.

Most graphics should go into the graphic_elements.py file, whose methods
should return either a Frame, which can then be added to the master
window. Or they should return a Frame and a number of it's children, in
a list, if the children will need to be updated at runtime.