Save File
Added Save file function, as well as a button which save the file.  The
open and save buttons are not in their final places.

Also began to implement a createMessageArea function, which will create
an area at the bottom of the master window dedicated to displaying
messages to the user ("file saved", error messages, etc.).