Send Message to User
Finished implementing the createMessageArea function in the
graphic_elements.py file, and created the neccessary function in the
main file to send messages to the user. The label which lies in the
message area frame is saved into the self.message_textbox variable,
which is used by the self.sendMessage function to update the text which
is found in the message area. Currently there is no way of clearing the
message until there is another message send to the user.