Object Oriented
Making this alot smarter. Added a LinkNotebook class, which will hold
all of the tabs for the application. Added a LinkTab class, which will
be the tabs for each part of the application. This class has it's own
default menu built in which is displayed along the top of the Frame, and
the constructor has an optional menu_dict argument for passing a menu to
that particular tab. This will be useful to have the default tab (which
will contain a text file) have the usual menu items like save, open,
etc., and other tabs have controls which do diffs, and other operations
needed.

Also, created a new module, logging.py, which contains the log function.
This should be used by the classes in this application so that changing
where the log goes is very simple.