Keywords and Docs
Changed some of the constructors and methods to use **kwargs, so that
not all keywords must be specified each time that object is created or
method called.

Also added some descriptions of what the methods do, how to use them,
and a general description of the graphic_elements.py file at the top.