Sizing
Made it so that all elements can expand horizontally to fill the root
window. And made it so that the text area in a TextTab can expand
vertically.  This makes it so the text area fills most of the screen and
the menu and filename label are at the top in the first two rows of the
grid layout.

I also implemented a row_counter in the TextTab class, which is used
each time a graphical element is placed, this makes it so I don't have
to keep track of which row I am on at each step, I can simple call each
method to place a graphical element and the counter takes care of the
rest.