File Open
Began implementing the file open command.

In the LinkTab class, this simply tries to open a new file.  One thing
to change later is that this currently hangs the program until a file is
successfully opened, so if the user presses cancel, the dialog is
reopened until they select a file.

In the TextTab class, this calls the parent function, then inserts the
file's contents into the textbox.