Modified Flag
Added a * to the filename label when a file has been modified, but not
yet saved. To do this, I needed to create the filename_label variable,
which is now used as the label text, and the full file path is the
filename. These are separate so that the * is never used accidentally to
save the file.