Default Menus
Changed the default menu in the LinkTab class. Also added a default menu
to the TextTab class. The way this was implemented before was to define
the default menu for the TextTab in the LinkTab class, since that is
where I started putting the menu before wishing to add more to it.  This
soon created problems as some commands that I wanted to put there were
not in the parent class or the OperationTab class, so there was some
checks that I had to perform. This is better.