New Tab Dialog Structure
Added the body of the new tab dialog class, including buttons.  There is
still a white space at the bottom which I need to figure out how to get
rid of. I also need to set up the ok button, and figure out a way to get
the values obtained from the user to the newTab method of the
LinkNotebook class.