New Tab Dialog
Inserted much of the parent's constructor into the NewTabDialog class's
__init__ method, so I could override the padding values that were being
used. This is because I have set a background color, and the padding was
making it so the background color of the root window was shown. For some
reason I could not reset the root window background, so here we are.