Operation Tab
Started creating the operation tab. Also made it so the addTab function
of the Link class can be passed the arguments 'kind' and 'label', if
these are not passed, then the dialog will be raised. This was added so
that I could write a script which automatically opens a new operations
tab so I could quickly see the layout.

In the future I would like every operation that the user can do to be
programmable, that way the user can create their own 'flows' to quickly
perform tasks.