Open File fix
While I was linting, the linter suggested that I change the line with
the open() function to use a with statement, this caused the file handle
to be closed before it could be returned to the user, so I have changed
it back.  This could perhaps be changed again to a with statement, but
simply return the file handle before the with statment terminates.