Differences
Added the difference method to the fops file. This takes two file names,
and finds all the lines in the first which do not appear in the second
file. This is difference than a diff, which compares files line by line.
Also, this seems to me much quicker than the original bash script that I
created, which is nice.

In this method, I have included an **opts parameter, which is currently
unused, but in the future will be used to pass optional parameters to
the method, to do things like: return the matches as well as the
nonmatches, case sensitivity, etc.