Destroy then Create
Fixed issue where the contol frame was being overlaid instead of
replaced. This does however make it so if the user switches tabs to look
at something while making a desicion about which tab to pick in the
dropdown menu, the menu is reset, which is not a major issue, but
something to think about.