Differences from Ops Tab
Added two Comboboxes to the Operation Tab controls so that users can
take diffs of files. These Comboboxes currently show a list of the
filename_labeles that are kept by each tab.  Ideally in the future this
will be changed so that the values shown are the names of each tab, and
then in the take_difference method, the actually filenames associated to
each tab would be used.