Tab Lables
Changed the Comboboxes to show the tab labels instead of the filename
labels for each tab. This required that the LinkNotebook class also
keeps track of the labels as well as a list of the tabs.

Ideally, each tab would keep track of it's own label, then instead of
the LinkNotebook class keeping track of two lists, and there is no
possibility of the lists not being in the same order, since the current
implementation relies on this.