Edit Points and Cancel button
Fixed the behavior of the Cancel button on the Cut Dialog page, which
now simple closes the dialog with no errors. Also added a jump point
before making any cuts, so the user can jump back. Currently there is a
second jump point being added by (I think) the delete function, so to
undo a cut, two undos must be done. I hope to fix this in the future.