Cut Debugging and Error Dialogs
The cut function from the fops module needed to be debugged. It turns
out that the files that I was testing on did not have a uniform number
of records in each line, so it was causing problems.  Now when the cut
method is called, a method checks to see what is the minimum number of
records a line has, then if the range passed by the user exceeds that,
nothing in the file is changed.