Field Separators
Added to the CutDialog so the user can specify what the field separator
should be. This allows for more than just csv files, and actually allows
the user to strip quotes in a very specific circumstance where the lines
of the file take the form

"Carlson,Braden"

by using cut(f="2", fs='"').